Add incomplete functionality of post a tweet
This is incomplete and buggy, when you enter event_detail and press
the "Send Twitter" button it will return an error of
"The request is understood, but it has been refused. An
accompanying error message will explain why. This code is used
when requests are being denied due to update limits. You currently
have Essential access which includes access to Twitter API v2
endpoints only. If you need access to this endpoint, you’ll need to
apply for Elevated access via the Developer Portal.